DRY; `NewStaticBuidler()`
