+++ daemon.go
@@ -41,7 +41,6 @@ import (
 	"google.golang.org/grpc/credentials"
 	"google.golang.org/grpc/credentials/insecure"
 	"google.golang.org/grpc/keepalive"
-	"google.golang.org/grpc/resolver"
 	"google.golang.org/protobuf/encoding/protojson"
 )
 
@@ -433,7 +432,7 @@ func (s *Daemon) Client() (V1Client, error) {
 
 	conn, err := grpc.DialContext(context.Background(),
 		fmt.Sprintf("static:///%s", s.PeerInfo.GRPCAddress),
-		grpc.WithResolvers(newStaticBuilder()),
+		grpc.WithResolvers(NewStaticBuilder()),
 		grpc.WithTransportCredentials(insecure.NewCredentials()))
 	if err != nil {
 		return nil, err
@@ -482,41 +481,3 @@ func WaitForConnect(ctx context.Context, addresses []string) error {
 	}
 	return nil
 }
-
-type staticBuilder struct{}
-
-var _ resolver.Builder = (*staticBuilder)(nil)
-
-func (sb *staticBuilder) Scheme() string {
-	return "static"
-}
-
-func (sb *staticBuilder) Build(target resolver.Target, cc resolver.ClientConn, _ resolver.BuildOptions) (resolver.Resolver, error) {
-	var resolverAddrs []resolver.Address
-	for _, address := range strings.Split(target.Endpoint(), ",") {
-		resolverAddrs = append(resolverAddrs, resolver.Address{
-			Addr:       address,
-			ServerName: address,
-		})
-	}
-	if err := cc.UpdateState(resolver.State{Addresses: resolverAddrs}); err != nil {
-		return nil, err
-	}
-	return &staticResolver{cc: cc}, nil
-}
-
-// newStaticBuilder returns a builder which returns a staticResolver that tells GRPC
-// to connect a specific peer in the cluster.
-func newStaticBuilder() resolver.Builder {
-	return &staticBuilder{}
-}
-
-type staticResolver struct {
-	cc resolver.ClientConn
-}
-
-func (sr *staticResolver) ResolveNow(_ resolver.ResolveNowOptions) {}
-
-func (sr *staticResolver) Close() {}
-
-var _ resolver.Resolver = (*staticResolver)(nil)
